Fix detached head issues by creating branch
This is a small hack to address the current issue with go-git noted in https://github.com/src-d/go-git/issues/707.

TLDR is that go-git doesn't have any mechanism for leaving git in the correct branch when dealing with a detached head, so without this fix a finished call of make sync or make validate will drop you back in the master (or default) branch.

This causes a huge issue since it breaks the fundamental assumption that sync makes that it is able to use git features to "reset" the index of assets it has not changed yet.

Signed-off-by: Arvind Iyengar <arvind.iyengar@rancher.com>