Add make release to staging template
This commit adds the ability to run `make release` to the Staging branch by creating a new script, updating the Makefile for the Staging branch to point to it, and creating a new release/README.md.

We ensure these new files only get added to the Staging branch by updating the update.yaml file.

Signed-off-by: Arvind Iyengar <arvind.iyengar@rancher.com>