Ignore package when storing charts and assets
Previously, we would follow the following scheme for the charts/ directory and assets/ directory:

```
charts/
  packageName/
    chartName/
      version/

assets/
  packageName/
    chartName-version.tgz
```

Instead, now we are switching to a package-less model for storing charts and assets.

```
charts/
  chartName/
    version/

assets/
  chartName/
    chartName-version.tgz
```

This allows us to ensure that multiple packages tracking the same underlying chart end up putting assets in the same directory.