Introduce make zip+unzip and refactor make charts
This commit introduces logic on comparing the contents of tgz files that will be used to power `make charts` and a new command called `make zip`.

Using this, `make charts` will now no longer require the use of Git to determine whether a chart is up-to-date; it can now directly do a deep check of the tgz files to determine whether to modify the tgz file. It will only modify an existing tgz file if necessary (e.g. if the SHA of any file has changed).

This is also used to help create a `make zip` command, which will take the current contents of a chart under `charts` and rezip it under assets, if the contents have changed. This will be used in order to allow modifying annotations of already released charts.

`make unzip`, similarly, will instead take the current contents of an asset under `assets` and unzip it into charts. This will allow debugging assets to be a much easier process.