Add make standardize
`make standardize` is a new command that will allow us to take any arbitrary Helm repository (defined as one with a `charts` directory tracking Helm charts) and convert it into one that we can directly start using with these scripts.

I tested the use of this function on rancher/charts and rancher/system-charts and it worked as expected.

The primary reason for exposing this as a `make` function is because it will be used internally to run `make validate` anyways; so we might as well make it available for use.