Ensure patches on Chart.yaml are standardized
This ensures that we always standardize the Chart.yaml on prepare and patch to ensure that patches that are generated don't look verbose due to switching the order of entries.

It also prevents the Chart.yaml from containing comments since those won't end up in the packaged chart either way.

It also adds experimental support for automatically adding a `Chart.yaml` if upstream does not have it, which enables us to declare manifest-based upstreams instead of just Helm-chart-based upstreams.