adding compareReleasedAndDevAssets function
This is where the most important comparisons happen by comparing the released and in devlopment assets versions.
We have 4 possibilities here...
released and in lifecycle: OK (nothing to be done).
not released and out lifecycle: OK (nothing to be done).
not released and in lifecycle: OK but needs to be addressed.
released and out lifecycle: ERROR this should never happen.

The function is just creating 4 different maps and appending to them according
to the previous rules above.