adding parsePackageYaml method
This method will enforce that the package.yaml for an auto-chart-bump is following the standards.
The standards will be defined at the rancher/charts/wiki.
They were already discussed with the chart owner (rancher-cis-operator).
They will serve as a standard for all other chart owners that desire to adopt the feature