Making new standard to deal with yaml files
There was a lot of duplicate functions and 2 different yaml pkg
versions to deal with.

The "gopkg.in/yaml.v2 and .v3" are not being maintained anymore but is
still the best package to handle yaml files in go.

The .v3 is more strict and performant to handle yaml files, however
there is a lot of old charts that will throw errors.

I am changing all the yaml file handling to these functions, only
allowing to be ignored already released old charts. All new charts will
not have the format ignored.